src/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
matplotlib
python-dotenv
PyMySQL
networkx
Werkzeug
reportlab
pillow
